fix: clearing interval, moving funcs and files. roi arr
