+++ vue-client/src/utils/overlap.js
@@ -1,6 +1,142 @@
-import uniqid from 'uniqid'
-import labels from './labels.json'
-import { xywh2xyxy } from './canvas_utility'
+// import uniqid from 'uniqid'
+const labels = [
+  'human',
+  'bicycle',
+  'car',
+  'motorcycle',
+  'airplane',
+  'bus',
+  'train',
+  'truck',
+  'boat',
+  'traffic light',
+  'fire hydrant',
+  'stop sign',
+  'parking meter',
+  'bench',
+  'bird',
+  'cat',
+  'dog',
+  'horse',
+  'sheep',
+  'cow',
+  'elephant',
+  'bear',
+  'zebra',
+  'giraffe',
+  'backpack',
+  'umbrella',
+  'handbag',
+  'tie',
+  'suitcase',
+  'frisbee',
+  'skis',
+  'snowboard',
+  'sports ball',
+  'kite',
+  'baseball bat',
+  'baseball glove',
+  'skateboard',
+  'surfboard',
+  'tennis racket',
+  'bottle',
+  'wine glass',
+  'cup',
+  'fork',
+  'knife',
+  'spoon',
+  'bowl',
+  'banana',
+  'apple',
+  'sandwich',
+  'orange',
+  'broccoli',
+  'carrot',
+  'hot dog',
+  'pizza',
+  'donut',
+  'cake',
+  'chair',
+  'couch',
+  'potted plant',
+  'bed',
+  'dining table',
+  'toilet',
+  'tv',
+  'laptop',
+  'mouse',
+  'remote',
+  'keyboard',
+  'cell phone',
+  'microwave',
+  'oven',
+  'toaster',
+  'sink',
+  'refrigerator',
+  'book',
+  'clock',
+  'vase',
+  'scissors',
+  'teddy bear',
+  'hair drier',
+  'toothbrush'
+]
+import { xywh2xyxy } from './canvas.js'
+
+export function detectionsToROIArr(
+  detections,
+  imageWidth,
+  imageHeight,
+  vehicleOnly
+) {
+  let condition = false;
+
+  let _predictionsArr = [];
+  const boxes = shortenedCol(detections, [0, 1, 2, 3]);
+  const scores = shortenedCol(detections, [4]);
+  const class_detect = shortenedCol(detections, [5]);
+  if (detections === undefined || detections.length <= 0) {
+    return [];
+  }
+
+  for (let index = 0; index < detections.length; index++) {
+    const detectionScore = scores[index];
+    const detectionClass = class_detect[index];
+    let dect_label = labels[detectionClass];
+    if (vehicleOnly === false) {
+      condition = true;
+    } else {
+      condition = isVehicle(dect_label);
+    }
+    if (condition) {
+      const roiObj = { cords: {} };
+      let [x1, y1, x2, y2] = xywh2xyxy(boxes[index]);
+      // Extract the bounding box coordinates from the 'boxes' tensor
+      y1 = y1 * (imageHeight / 640);
+      y2 = y2 * (imageHeight / 640);
+      x1 = x1 * (imageWidth / 640);
+      x2 = x2 * (imageWidth / 640);
+      let dect_width = x2 - x1;
+      let dect_weight = y2 - y1;
+      roiObj.cords.bottom_y = y2;
+      roiObj.cords.left_x = x2;
+      roiObj.cords.top_y = y1;
+      roiObj.cords.right_x = x1;
+      roiObj.cords.width = dect_width;
+      roiObj.cords.height = dect_weight;
+      // Add the detection score to the bbox object
+      roiObj.confidenceLevel = detectionScore;
+      roiObj.label = dect_label;
+      roiObj.area = dect_width * dect_weight;
+      // Add the bbox object to the bboxes array
+
+      _predictionsArr.push(roiObj);
+    }
+  }
+
+  return _predictionsArr;
+}
+
 
 //The getOverlap function takes two rectangles as input and
 // returns a new rectangle that represents the overlapping area between the two rectangles.
@@ -100,11 +236,14 @@ export function isVehicle(label) {
 
 export function selectedFactory(cords) {
   let date = new Date()
+  const uuid = 'xxxxxxxx-xxxx-4xxx-yxxx'.replace(/[x]/g, () =>
+    ((Math.random() * 16) | 0).toString(16)
+  )
 
   const roiObj = {
+    uuid,
     label: 'vehicle',
     cords: { ...cords },
-    uid: uniqid() + 'DATE' + Date.now().toString(),
     area: Math.round(cords.width * cords.height),
     firstSeen: null,
     lastSeen: null,
@@ -138,7 +277,7 @@ export function totalOccupied(roiArr) {
   return { OccupiedCount, availableCount }
 }
 
-export function detectionsToROIArr(detections, imageWidth, imageHeight, vehicleOnly) {
+export function detectionsToRoi(detections, imageWidth, imageHeight, vehicleOnly) {
   let condition = false
 
   let _predictionsArr = []
